GraphML

PowerModelsONM.add_graphml_data! — Method

add_graphml_data!(node::EzXML.Node, key::String, value::Any)::EzXML.Node

Helper function to add an AttributeNode with key and value to a node

PowerModelsONM.add_root_graphml_node! — Method

add_root_graphml_node!(doc::EzXML.Document)::EzXML.Node

Helper function to build 'graphml' root XML Node for GraphML XML documents

PowerModelsONM.build_graphml_document — Method

build_graphml_document(eng::Dict{String,<:Any}; type::Type="nested")

Helper function to build GraphML XML document from a eng network data structure.

type controls whether the resulting graph is a NestedGraph, i.e., buses are contained within load blocks, or a UnnestedGraph, where node groups are not utilized.

PowerModelsONM.build_graphml_document — Method

build_graphml_document(gr::NestedGraph)::EzXML.Document

Helper function to build GraphML XML document from a NestedGraph

PowerModelsONM.build_graphml_document — Method

build_graphml_document(gr::UnnestedGraph)::EzXML.Document

Helper function to build GraphML XML document from an UnnestedGraph

PowerModelsONM.build_graphml_edge — Method

build_graphml_edge(id::String, source::String, target::String)::EzXML.Node

Helper function to build an 'edge' XML Node object for GraphML XML documents

PowerModelsONM.build_graphml_graph — Function

build_graphml_graph(id::String, directed::Bool=false)::EzXML.Node

Helper function to build a 'graph' XML Node for GraphML XML documents

PowerModelsONM.build_graphml_key — Function

build_graphml_key(id::String, is_for::String, attr_name::String, attr_type::String, default::Any=missing)::EzXML.Node

Helper function to build an XML AttributeNode for attribute data for GraphML XML documents

PowerModelsONM.build_graphml_node — Method

build_graphml_node(id::String)::EzXML.Node

Helper function to build graph 'node' XML Node for GraphML XML documents

PowerModelsONM.build_nested_graph — Method

build_nested_graph(eng::Dict{String,Any})::NestedGraph

Helper function to build a NestedGraph of network data eng

PowerModelsONM.build_unnested_graph — Method

build_unnested_graph(eng::Dict{String,<:Any})::UnnestedGraph

Helper function to build an UnnestedGraph from eng network data.

PowerModelsONM.save_graphml — Method

save_graphml(io::IO, eng::Dict{String,<:Any}; type::String="nested")

Save a GraphML XML document built from eng network data to IO stream.

type controls whether the resulting graph is a NestedGraph, i.e., buses are contained within load blocks, or a UnnestedGraph, where node groups are not utilized.

PowerModelsONM.save_graphml — Method

save_graphml(graphml_file::String, eng::Dict{String,<:Any}; type::String="nested")

Save a GraphML XML document built from eng network data to graphml_file.

type controls whether the resulting graph is a NestedGraph, i.e., buses are contained within load blocks, or a UnnestedGraph, where node groups are not utilized.