IO

Parsers

PowerModelsONM.parse_network — Function

parse_network(network_file::String)::Tuple{Dict{String,Any},Dict{String,Any}}

Parses network file given by runtime arguments into its base network, i.e., not expanded into a multinetwork, and multinetwork, which is the multinetwork ENGINEERING representation of the network.

PowerModelsONM.parse_events — Function

parse_events(events_file::String; validate::Bool=true)::Vector{Dict{String,Any}}

Parses an events file into a raw events data structure

Validation

If validate=true (default), the parsed data structure will be validated against the latest Events Schema.

parse_events(raw_events::Vector{<:Dict{String,<:Any}}, mn_data::Dict{String,<:Any})::Dict{String,Any}

Converts raw_events, e.g. loaded from JSON, and therefore in the format Vector{Dict}, to an internal data structure that closely matches the multinetwork data structure for easy merging (applying) to the multinetwork data structure.

Will attempt to find the correct subnetwork from the specified timestep by using "mn_lookup" in the multinetwork data structure.

Switch events

Will find the correct switch id from a source_id, i.e., the assettype.name from the source file, which for switches will be line.name, and create a data structure containing the properties defined in `eventdata` under the native ENGINEERING switch id.

Fault events

Will attempt to find the appropriate switches that need to be OPEN to isolate a fault, and disable them, i.e., set dispatchable=false, until the end of the duration of the fault, which is specified in milliseconds.

It will re-enable the switches, i.e., set dispatchable=true after the fault has ended, if the next timestep exists, but will not automatically set the switches to CLOSED again; this is a decision for the algorithm optimize_switches to make.

parse_events(events_file::String, mn_data::Dict{String,<:Any}; validate::Bool=true)::Dict{String,Any}

Parses raw events from events_file and passes it to parse_events to convert to the native data type.

Validation

If validate=true (default), the parsed data structure will be validated against the latest Events Schema.

PowerModelsONM.parse_settings — Function

parse_settings(settings_file::String; validate::Bool=true)::Dict{String,Any}

Parses network settings JSON file.

Validation

If validate=true (default), the parsed data structure will be validated against the latest Settings Schema.

PowerModelsONM.parse_faults — Function

parse_faults(faults_file::String; validate::Bool=true)::Dict{String,Any}

Parses fault JSON input files which have the same structure as the outputs from PowerModelsProtection.build_mc_fault_stuides

Expected JSON Structure

{
    "bus_name": {
        "fault_type": {
            "fault_key": {
                "g": [ [200, -100, -100], [-100, 200, -100], [-100, -100, 200]],
                "b": [ [0, 0, 0], [0, 0, 0], [0, 0, 0]],
                "status": "ENABLED",
                "fault_type": "fault_type",
                "bus": "bus_name",
                "name": "fault_key",
                "connections": [1, 2, 3]
            }
        }
    }
}

where "fault_type" is one of:

"3p" : 3-phase

"3pg" : 3-phase-to-ground

"ll" : line-to-line

"llg" : line-to-line-to-ground

"ll" : line-to-line

"bus_name" is arbitrary, and just needs to match a bus's name in the network model.

"status" is a PowerModelsDistribution.Status Enum in String form, and must be either "ENABLED" or "DISABLED".

"g" and "b" are matrices in SI units.

"fault_type" in the deepest level is merely metadata and should match the "fault_type" key above.

"name" should match the fault key and is required to be an Integer.

"connections" is a Vector of Integers indicating the phases that the fault applies to.

For more details see PowerModelsProtection's documentation

Validation

If validate=true (default), the parsed data structure will be validated against the latest Faults Schema.

PowerModelsONM.parse_inverters — Function

parse_inverters(inverter_file::String; validate::Bool=true)::Dict{String,Any}

Parses an inverters JSON file, used in run_stability_analysis!

Validation

If validate=true (default), the parsed data structure will be validated against the latest Inverters Schema.

Applicators

PowerModelsONM.apply_events — Function

apply_events(network::Dict{String,<:Any}, events::Dict{String,<:Any})::Dict{String,Any}

Creates a copy of the multinetwork data structure network and applies the events in events to that data.

PowerModelsONM.apply_events! — Function

apply_events!(args::Dict{String,<:Any})::Dict{String,Any}

Applies events in-place using apply_events, for use inside of entrypoint

PowerModelsONM.apply_settings — Function

apply_settings(network::Dict{String,<:Any}, settings::Dict{String,<:Any})::Dict{String,Any}

Applies settings to multinetwork network

PowerModelsONM.apply_settings! — Function

apply_settings!(args::Dict{String,Any})::Dict{String,Any}

Applies settings to the network

PowerModelsONM.initialize_output — Function

initialize_output(args::Dict{String,<:Any})::Dict{String,Any}

Initializes the empty data structure for "output_data"

Writers

PowerModelsONM.write_json — Function

write_json(file::String, data::Dict{String,<:Any}; indent::Union{Int,Missing}=missing)

Write JSON data to file. If !ismissing(indent), JSON will be pretty-formatted with indent

JSON Schema

PowerModelsONM.load_schema — Function

load_schema(file::String)::JSONSchema.Schema

Loads a JSON Schema for validation, fixing the ref paths inside the schemas on load

PowerModelsONM.validate_events — Method

validate_events(data::Vector{Dict})::Bool

Validates events data against models/events schema

PowerModelsONM.validate_faults — Method

validate_faults(data::Dict)::Bool

Validates fault input data against models/faults schema

PowerModelsONM.validate_inverters — Method

validate_inverters(data::Dict)::Bool

Validates inverter data against models/inverters schema

PowerModelsONM.validate_output — Method

validate_output(data::Dict)::Bool

Validates output data against models/outputs schema

PowerModelsONM.validate_runtime_arguments — Method

validate_runtime_arguments(data::Dict)::Bool

Validates runtimearguments data against models/runtimearguments schema

PowerModelsONM.validate_settings — Method

validate_settings(data::Dict)::Bool

Validates runtimesettings data against models/runtimesettings schema