Optimization Problems

Optimal Switching / Maximal Load Delivery (MLD)

PowerModelsONM.optimize_switches — Function

optimize_switches(network::Dict{String,<:Any}, solver; formulation::Type=LPUBFSwitchPowerModel, algorithm::String="global")::Dict{String,Any}

Iterates over all subnetworks in a multinetwork data structure network, in order, and solves the optimal switching / MLD problem sequentially, updating the next timestep with the new switch configurations and storage energies from the solved timestep.

optimize_switches(subnetwork::Dict{String,<:Any}, prob::Function, solver; formulation=LPUBFSwitchPowerModel)::Dict{String,Any}

Optimizes switch states for load shedding on a single subnetwork (not a multinetwork), using prob

Optionally, a PowerModelsDistribution formulation can be set independently, but is LinDist3Flow by default.

PowerModelsONM.solve_mc_osw_mld_mi — Function

solve_mc_osw_mld_mi(data::Union{Dict{String,<:Any}, String}, model_type::Type, solver;
    solution_processors::Vector{Function}=Function[],
    eng2math_passthrough::Dict{String,Vector{String}}=Dict{String,Vector{String}}(),
    ref_extensions::Vector{Function}=Function[], kwargs...
)::Dict{String,Any}

Solves a multiconductor optimal switching (mixed-integer) problem using model_type and solver

Calls back to PowerModelsDistribution.solvemcmodel, and therefore will accept any valid kwargs for that function. See PowerModelsDistribution documentation for more details.

PowerModelsONM.solve_mn_mc_osw_mld_mi — Function

solve_mn_mc_osw_mld_mi(data::Union{Dict{String,<:Any}, String}, model_type::Type, solver;
    solution_processors::Vector{Function}=Function[],
    eng2math_passthrough::Dict{String,Vector{String}}=Dict{String,Vector{String}}(),
    ref_extensions::Vector{Function}=Function[], kwargs...
)::Dict{String,Any}

Solves a multinetwork multiconductor optimal switching (mixed-integer) problem using model_type and solver

Calls back to PowerModelsDistribution.solvemcmodel, and therefore will accept any valid kwargs for that function. See PowerModelsDistribution documentation for more details.

PowerModelsONM.build_mn_mc_opf_oltc_capc — Function

build_mn_mc_opf_oltc_capc(pm::AbstractUnbalancedPowerModel)

constructor for bus injection opf

build_mn_mc_opf_oltc_capc(pm::PMD.AbstractUBFModels)

constructor for branch flow opf

Optimal Dispatch

PowerModelsONM.optimize_dispatch — Function

optimize_dispatch(network::Dict{String,<:Any}, formulation::Type, solver; switching_solutions::Union{Missing,Dict{String,<:Any}}=missing)::Dict{String,Any}

Solve a multinetwork optimal power flow (solve_mn_mc_opf) using formulation and solver

Fault stuides

PowerModelsONM.run_fault_studies — Function

run_fault_studies(network::Dict{String,<:Any}, solver; faults::Dict{String,<:Any}=Dict{String,Any}(), switching_solutions::Union{Missing,Dict{String,<:Any}}=missing, dispatch_solution::Union{Missing,Dict{String,<:Any}}=missing, distributed::Bool=false)::Dict{String,Any}

Runs fault studies defined in ieee13faults.json. If no faults file is provided, it will automatically generate faults using `PowerModelsProtection.buildmcfaultstudy`.

It will convert storage to limited generators, since storage is not yet supported in IVRU models in PowerModelsProtection

Uses run_fault_study to solve the actual fault study.

solver will determine which instantiated solver is used, "nlp_solver" or "juniper_solver"

PowerModelsONM.run_fault_study — Function

run_fault_study(subnetwork::Dict{String,<:Any}, faults::Dict{String,<:Any}, solver)::Dict{String,Any}

Uses PowerModelsProtection.solve_mc_fault_study to solve multiple faults defined in faults, applied to subnetwork, i.e., not a multinetwork, using a nonlinear solver.

Requires the use of PowerModelsDistribution.IVRUPowerModel.

Stability Analysis

PowerModelsONM.run_stability_analysis — Function

run_stability_analysis(network, inverters::Dict{String,<:Any}, solver; formulation::Type=PMD.ACRUPowerModel, switching_solutions::Union{Missing,Dict{String,<:Any}}=missing, distributed::Bool=false)::Dict{String,Bool}

Runs small signal stability analysis using PowerModelsStability and determines if each timestep configuration is stable

inverters is an already parsed inverters file using parse_inverters

The formulation can be specified with formulation, but note that it must result in "vm" and "va" variables in the solution, or else PowerModelsDistribution.sol_data_model! must support converting the voltage variables into polar coordinates.

solver for stability analysis (NLP OPF)

run_stability_analysis(subnetwork::Dict{String,<:Any}, omega0::Real, rN::Int, solver; formulation::Type=PMD.ACPUPowerModel)::Bool

Runs stability analysis on a single subnetwork (not a multinetwork) using a nonlinear solver.

Optimization solvers

PowerModelsONM.build_solver_instances — Function

build_solver_instances(;
    nlp_solver::Union{Missing,JuMP.MOI.OptimizerWithAttributes}=missing,
    mip_solver::Union{Missing,JuMP.MOI.OptimizerWithAttributes}=missing,
    minlp_solver::Union{Missing,JuMP.MOI.OptimizerWithAttributes}=missing,
    misocp_solver::Union{Missing,JuMP.MOI.OptimizerWithAttributes}=missing,
    nlp_solver_options::Union{Missing,Vector{Pair}}=missing,
    mip_solver_options::Union{Missing,Vector{Pair}}=missing,
    minlp_solver_options::Union{Missing,Vector{Pair}}=missing,
    feas_tol::Float64=1e-6,
    mip_gap_tol::Float64=1e-4,
    gurobi::Bool=false,
    knitro::Bool=false,
    verbose::Bool=false,
    debug::Bool=false,
)::Dict{String,JuMP.MOI.OptimizerWithAttributes}

Returns solver instances as a Dict ready for use with JuMP Models, for NLP ("nlp_solver"), MIP ("mip_solver"), MINLP ("minlp_solver"), and (MI)SOC ("misocp_solver") problems.

nlp_solver (default: missing): If missing, will use Ipopt as NLP solver, or KNITRO if knitro=true

mip_solver (default: missing): If missing, will use Cbc as MIP solver, or Gurobi if gurobi==true

minlp_solver (default: missing): If missing, will use Juniper with nlp_solver and mip_solver, of KNITRO if knitro=true

misocp_solver (default: missing): If missing will use Juniper with mip_solver, or Gurobi if gurobi==true

nlp_solver_options (default: missing): If missing, will use some default nlp solver options

mip_solver_options (default: missing): If missing, will use some default mip solver options

minlp_solver_options (default: missing): If missing, will use some default minlp solver options

feas_tol (default: 1e-4): The solver tolerance

mip_gap_tol (default: 0.05): The desired MIP Gap for the MIP Solver

verbose (default: false): Sets the verbosity of the solvers

debug (default: false): Sets the verbosity of the solvers even higher (if available)

gurobi (default: false): Use Gurobi for MIP / MISOC solvers

knitro (default: false): Use KNITRO for NLP / MINLP solvers