Data

PowerModelsONM.convert — Function

convert(value::Any, path::Tuple{Vararg{String}}=tuple())

Helper function to assist in converting deprecated settings to their correct types / values

PowerModelsONM.delete_path! — Method

delete_path!(data::Dict{String,<:Any}, path::Tuple{Vararg{String}})

Helper function for variant where data is not a dict

PowerModelsONM.delete_path! — Method

delete_path!(data::Dict{String,<:Any}, path::Tuple{Vararg{String}})

Helper function to delete a path within a dictionary

PowerModelsONM.recursive_merge — Method

recursive other merge, with vectors getting overwritten instead of appended

PowerModelsONM.recursive_merge — Method

recursive dictionary merge, similar to update data, with vectors getting overwritten instead of appended

PowerModelsONM.recursive_merge_including_vectors — Method

recursive other merge

PowerModelsONM.recursive_merge_including_vectors — Method

recursive dictionary merge, similar to update data

PowerModelsONM.recursive_merge_including_vectors — Method

recursive vector merge, similar to update data

PowerModelsONM.recursive_merge_timesteps — Method

recursive_merge_timesteps(x::T, y::U)::promote_type(T,U) where {T<: AbstractVector,U<: AbstractVector}

helper function to recursively merge timestep vectors (e.g., of dictionaries)

PowerModelsONM.set_dict_value! — Method

set_dict_value!(a::Dict, key::String, value::Any)

Helper function to assist in setting nested Dict values

PowerModelsONM.set_dict_value! — Method

set_dict_value!(a::T, path::Tuple{Vararg{String}}, value::Any) where T <: Dict

Helper function to assist in setting nested Dict values