Data

Transformations

PowerModelsONM.convert_storage! — Method

convert_storage!(nw::Dict{String,Any})

Helper function for PowerModelsProtection fault studies to convert storage to generators in a subnetwork; PowerModelsProtection currently does not support storage object constraints / variables, so this is a workaround until those constraints/variables are added.

This works on ENGINEERING subnetworks (not multinetworks).

Data Handling

PowerModelsONM.are_blocks_warm — Method

are_blocks_warm(data::Dict{String,<:Any}, blocks)::Dict{Set,Bool}

Identifies whether load blocks, which is a set of set of buses, are "warm" or not using is_block_warm

PowerModelsONM.identify_cold_loads — Method

identify_cold_loads(data::Dict)::Dict{String,Bool}

Identifies whether loads are currently "cold" or not using are_blocks_warm

PowerModelsONM.is_block_warm — Method

is_block_warm(data::Dict, block::Set)::Bool

Returns true if a load block is "warm", i.e., has an active generation element within the block

Solution Handling

PowerModelsONM.apply_switch_solutions! — Method

apply_switch_solutions!(network::Dict{String,<:Any}, optimal_switching_results::Dict{String,<:Any})::Dict{String,Any}

Updates a multinetwork network in-place with the results from optimal switching optimal_switching_results.

Used when not using the in-place version of optimize_switches!.

PowerModelsONM.apply_switch_solutions — Method

apply_switch_solutions(network::Dict{String,<:Any}, optimal_switching_results::Dict{String,<:Any})::Dict{String,Any}

Creates a copy of the network with the solution copied in from optimal_switching_results