Data

Data Handling

PowerModelsONM.count_faults — Function

count_faults(faults::Dict{String,<:Any})::Int

Helper function to count the total number of faults.

Solution Handling

PowerModelsONM.apply_switch_solutions! — Method

apply_switch_solutions!(network::Dict{String,<:Any}, optimal_switching_results::Dict{String,<:Any})::Dict{String,Any}

Updates a multinetwork network in-place with the results from optimal switching optimal_switching_results.

Used when not using the in-place version of optimize_switches!.

PowerModelsONM.apply_switch_solutions — Method

apply_switch_solutions(network::Dict{String,<:Any}, optimal_switching_results::Dict{String,<:Any})::Dict{String,Any}

Creates a copy of the network with the solution copied in from optimal_switching_results.

PowerModelsONM.solution_blocks! — Method

solution_blocks!(pm::AbstractUnbalancedPowerModel, sol::Dict{String,Any})

Adds block ids (as generated in the ref), and microgrid_ids to the solution

PowerModelsONM.solution_inverter! — Method

solution_inverter!(pm::AbstractUnbalancedPowerModel, sol::Dict{String,Any})

Converts inverter to Inverter enum, across all time steps.

PowerModelsONM.solution_reference_buses! — Method

solution_reference_buses!(pm::AbstractUnbalancedPowerModel, sol::Dict{String,Any})

Raises bus_type from math model up to solution for reporting, across all time steps.

PowerModelsONM.solution_statuses! — Method

solution_statuses!(pm::AbstractUnbalancedPowerModel, sol::Dict{String,Any})

Converts all status fields in a solution sol from Float64 to Status enum, for all time steps.