Internal Functions

PowerModelsONM._apply_to_network! — Method

helper function that applies settings to the network objects of type

PowerModelsONM._convert_depreciated_runtime_args! — Method

helper function to convert depreciated runtime arguments to their appropriate network settings structure

PowerModelsONM._convert_event_data_types! — Method

helper function to convert JSON data types to native data types (Enums) in events

PowerModelsONM._convert_to_settings! — Method

converts depreciated global settings, e.g. voltage-lower-bound, to the proper way to specify settings

PowerModelsONM._convert_voltage_bound_to_settings! — Method

helper function to convert

PowerModelsONM._deepcopy_args! — Method

_deepcopy_args!(args::Dict{String,<:Any})::Dict{String,Any}

Copies arguments to "raw_args" in-place in args, for use in entrypoint

PowerModelsONM._find_next_nw_id_from_fault_duration — Method

helper function to find the next timestep following a fault given its duration in ms

PowerModelsONM._find_nw_id_from_timestep — Method

helper function to find the multinetwork id of the subnetwork corresponding most closely to a timestep

PowerModelsONM._find_switch_id_from_source_id — Method

helper function to find a switch id in the network model based on the dss source_id

PowerModelsONM._find_switch_ids_by_faulted_asset — Method

helper function to find which switches need to be opened to isolate a fault on asset given by source_id

PowerModelsONM._fix_fault_data_types! — Method

helper function to help parse data types for multiconductor fault study data structures from JSON

PowerModelsONM._get_dispatch_formulation — Method

helper function to convert from opt-disp-formulation string to PowerModelsDistribution Type

PowerModelsONM._get_switch_formulation — Method

helper function to convert from opt-switch-formulation string to PowerModelsONM Type

PowerModelsONM._make_filtered_logger — Method

_make_filtered_logger(mods::Vector, level::Logging.LogLevel)

Helper function to create the filtered logger for PMD

PowerModelsONM._prepare_dispatch_data — Function

prepares data for running a optimal dispatch problem, copying in solutions from the switching results, if present

PowerModelsONM._prepare_fault_study_multinetwork_data — Function

helper function that helps to prepare all of the subnetworks for use in PowerModelsProtection.solve_mc_fault_study

PowerModelsONM._prepare_optimal_switching_data — Method

helper function to prepare optimal switching data structure

PowerModelsONM._prepare_stability_multinetwork_data — Function

helper function to prepare the multinetwork data for stability analysis (adds inverters, data_model)

PowerModelsONM._ref_add_load_blocks! — Method

Ref extension to add load blocks to ref

PowerModelsONM._ref_add_max_switch_actions! — Method

Ref extension to add maxswitchactions to ref, and set to Inf if option is missing

PowerModelsONM._sanitize_results_metadata! — Method

turns any fields that is not a number into a string

PowerModelsONM._update_storage_capacity! — Method

helper function to update storage capacity for the next subnetwork based on a solution

PowerModelsONM._update_switch_settings! — Method

helper function to update switch settings from a solution

PowerModelsONM._validate_against_schema — Method

validates dict or vector structure against json schema using JSONSchema.jl

PowerModelsONM._validate_against_schema — Method

validates dict or vector structure against json schema given by schema_name