Optimization Problems

Optimal Switching / Maximal Load Delivery (MLD)

PowerModelsONM.optimize_switches — Function

optimize_switches(network::Dict{String,<:Any}, solver; gurobi::Bool=false)::Dict{String,Any}

Iterates over all subnetworks in a multinetwork data structure network, in order, and solves the optimal switching / MLD problem sequentially, updating the next timestep with the new switch configurations and storage energies from the solved timestep.

optimize_switches(subnetwork::Dict{String,<:Any}, prob::Function, solver; formulation=PMD.LPUBFDiagPowerModel)::Dict{String,Any}

Optimizes switch states for load shedding on a single subnetwork (not a multinetwork), using prob (solve_mc_osw_mld_mi_indicator, if you are using a solver that supports indicator constraints like Gurobi or CPLEX, or solve_mc_osw_mld_mi), solver.

Optionally, a PowerModelsDistribution formulation can be set independently, but is LinDist3Flow by default.

PowerModelsONM.solve_mc_osw_mld_mi_indicator — Function

solve_mc_osw_mld_mi_indicator(data::Union{Dict{String,<:Any}, String}, model_type::Type, solver; kwargs...)::Dict{String,Any}

Solves a multiconductor optimal switching (mixed-integer) problem using model_type and solver

Uses indicator constraints, so requires a solver that supports them (e.g., Gurobi or CPLEX)

Calls back to PowerModelsDistribution.solvemcmodel, and therefore will accept any valid kwargs for that function. See PowerModelsDistribution documentation for more details.

PowerModelsONM.solve_mc_osw_mld_mi — Function

solve_mc_osw_mld_mi(data::Union{Dict{String,<:Any}, String}, model_type::Type, solver; kwargs...)::Dict{String,Any}

Solves a multiconductor optimal switching (relaxed, i.e., not mixed-integer) problem using model_type and solver

Calls back to PowerModelsDistribution.solvemcmodel, and therefore will accept any valid kwargs for that function. See PowerModelsDistribution documentation for more details.

PowerModelsONM.solve_mn_mc_osw_mld_mi_indicator — Function

solve_mn_mc_osw_mld_mi_indicator(data::Union{Dict{String,<:Any}, String}, model_type::Type, solver; kwargs...)::Dict{String,Any}

Solves a multinetwork multiconductor optimal switching (mixed-integer) problem using model_type and solver

Uses indicator constraints, so requires a solver that supports them (e.g., Gurobi or CPLEX)

Calls back to PowerModelsDistribution.solvemcmodel, and therefore will accept any valid kwargs for that function. See PowerModelsDistribution documentation for more details.

Optimal Dispatch

PowerModelsONM.optimize_dispatch — Function

optimize_dispatch(network::Dict{String,<:Any}, formulation::Type, solver)::Dict{String,Any}

Solve a multinetwork optimal power flow (solve_mn_mc_opf) using formulation and solver

Fault stuides

PowerModelsONM.run_fault_studies — Function

run_fault_studies(network::Dict{String,<:Any}, solver; faults::Dict{String,<:Any}=Dict{String,Any}())::Dict{String,Any}

Runs fault studies defined in faults.json. If no faults file is provided, it will automatically generate faults using PowerModelsProtection.build_mc_fault_study.

It will convert storage to limited generators, since storage is not yet supported in IVRU models in PowerModelsProtection

Uses run_fault_study to solve the actual fault study.

solver will determine which instantiated solver is used, "nlp_solver" or "juniper_solver"

PowerModelsONM.run_fault_study — Function

run_fault_study(subnetwork::Dict{String,<:Any}, faults::Dict{String,<:Any}, solver)::Dict{String,Any}

Uses PowerModelsProtection.solve_mc_fault_study to solve multiple faults defined in faults, applied to subnetwork, i.e., not a multinetwork, using a nonlinear solver.

Requires the use of PowerModelsDistribution.IVRUPowerModel.

Stability Analysis

PowerModelsONM.run_stability_analysis — Function

run_stability_analysis(network, inverters::Dict{String,<:Any}, solver; formulation::Type=PMD.ACRUPowerModel)::Dict{String,Bool}

Runs small signal stability analysis using PowerModelsStability and determines if each timestep configuration is stable

inverters is an already parsed inverters file using parse_inverters

The formulation can be specified with formulation, but note that it must result in "vm" and "va" variables in the solution, or else PowerModelsDistribution.sol_data_model! must support converting the voltage variables into polar coordinates.

solver for stability analysis (NLP OPF)

run_stability_analysis(subnetwork::Dict{String,<:Any}, omega0::Real, rN::Int, solver; formulation::Type=PMD.ACRUPowerModel)::Bool

Runs stability analysis on a single subnetwork (not a multinetwork) using a nonlinear solver.

Optimization solvers

PowerModelsONM.build_solver_instances — Function

build_solver_instances(; nlp_solver=missing, nlp_solver_tol::Real=1e-4, mip_solver=missing, mip_solver_tol::Real=1e-4, minlp_solver=missing, misocp_solver=missing, gurobi::Bool=false, verbose::Bool=false, debug::Bool=false)::Dict{String,Any}

Returns solver instances as a Dict ready for use with JuMP Models, for NLP ("nlp_solver"), MIP ("mip_solver"), MINLP ("minlp_solver"), and (MI)SOC ("misocp_solver") problems.

nlp_solver (default: missing): If missing, will use Ipopt as NLP solver

nlp_solver_tol (default: 1e-4): The solver tolerance

mip_solver (default: missing): If missing, will use Cbc as MIP solver, or Gurobi if gurobi==true

minlp_solver (default: missing): If missing, will use Alpine with nlp_solver and mip_solver

misocp_solver (default: missing): If missing will use Juniper with mip_solver, or Gurobi if gurobi==true

verbose (default: false): Sets the verbosity of the solvers

debug (default: false): Sets the verbosity of the solvers even higher (if available)

gurobi (default: false): Use Gurobi for MIP / MISOC solvers