Schema

Validation

PowerModelsONM.check_switch_state_feasibility — Method

check_switch_state_feasibility(data::Dict{String,<:Any})::Union{Dict{String,Bool},Bool}

Helper function to aid users in determining whether network model has a feasible starting switch configuration (at each time step, if the network model is multinetwork), assuming radiality constraints are applied.

PowerModelsONM.evaluate_events — Method

evaluate_events(data::Dict)

Helper function to give detailed output on JSON Schema validation of events data

PowerModelsONM.evaluate_output — Method

evaluate_output(data::Dict)

Helper function to give detailed output on JSON Schema validation of output data

PowerModelsONM.evaluate_robust_partitions — Method

evaluate_robust_partitions(data::Dict)

Helper function to give detailed output on JSON Schema validation of settings data

PowerModelsONM.evaluate_runtime_arguments — Method

evaluate_runtime_arguments(data::Dict)

Helper function to give detailed output on JSON Schema validation of runtime arguments data

PowerModelsONM.evaluate_settings — Method

evaluate_settings(data::Dict)

Helper function to give detailed output on JSON Schema validation of settings data

PowerModelsONM.validate_events — Method

validate_events(data::Vector{Dict})::Bool

Validates events data against models/events schema

PowerModelsONM.validate_faults — Method

validate_faults(data::Dict)::Bool

Validates fault input data against models/faults schema

PowerModelsONM.validate_inverters — Method

validate_inverters(data::Dict)::Bool

Validates inverter data against models/inverters schema

PowerModelsONM.validate_output — Method

validate_output(data::Dict)::Bool

Validates output data against models/outputs schema

PowerModelsONM.validate_robust_partitions — Method

validate_robust_partitions(data::Vector{Dict})::Bool

Validates events data against robust partitions output schema

PowerModelsONM.validate_runtime_arguments — Method

validate_runtime_arguments(data::Dict)::Bool

Validates runtimearguments data against models/runtimearguments schema

PowerModelsONM.validate_settings — Method

validate_settings(data::Dict)::Bool

Validates runtimesettings data against models/runtimesettings schema