Variables and Constraints

Variables

PowerModelsONM.variable_mc_load_block_indicator — Method

variable_mc_load_block_indicator(pm::PMD.AbstractUnbalancedPowerModel; nw::Int=PMD.nw_id_default, relax::Bool=false, report::Bool=true)

create variables for demand status by load block

Constraints

PowerModelsONM.constraint_load_block_isolation — Method

constraint_load_block_isolation(pm::PMD.AbstractUnbalancedPowerModel; nw::Int=PMD.nw_id_default, relax::Bool=true)

constraint to ensure that load blocks are properly isolated by opening switches

PowerModelsONM.constraint_switch_state_max_actions — Method

constraint_switch_state_max_actions(pm::PMD.AbstractUnbalancedPowerModel; nw::Int=PMD.nw_id_default)

max switching actions per timestep constraint

PowerModelsONM.constraint_load_block_isolation — Method

constraint_load_block_isolation(pm::PMD.AbstractUnbalancedPowerModel, nw::Int; relax::Bool=true)

constraint to ensure that load blocks get properly isolated by opening switches

\[\begin{align} & z^{sw}_i \leq z^d_b\ \forall i \in S,\forall b \in L \\ & z^{sw}_i \geq 0\ \forall i \in S \end{align}\]

PowerModelsONM.constraint_switch_state_max_actions — Method

constraint_switch_state_max_actions(pm::PMD.AbstractUnbalancedPowerModel, nw::Int)

max actions per timestep switch constraint

\[\sum_{\substack{i\in S}}{\Delta^{sw}_i} \leq z^{swu}\]

Objectives

PowerModelsONM.objective_mc_min_load_setpoint_delta_switch — Method

objective_mc_min_load_setpoint_delta_switch(pm::PMD.AbstractUnbalancedPowerModel)

minimum load delta objective (continuous load shed) with storage

\[\begin{align} \mbox{minimize: } & \nonumber \\ & \sum_{\substack{i\in N,c\in C}}{10 \left (1-z^v_i \right )} + \nonumber \\ & \sum_{\substack{i\in L,c\in C}}{10 \omega_{i,c}\left |\Re{\left (S^d_i\right )}\right |\left ( 1-z^d_i \right ) } + \nonumber \\ & \sum_{\substack{i\in H,c\in C}}{\left | \Re{\left (S^s_i \right )}\right | \left (1-z^s_i \right ) } + \nonumber \\ & \sum_{\substack{i\in G,c\in C}}{\Delta^g_i } + \nonumber \\ & \sum_{\substack{i\in B,c\in C}}{\Delta^b_i} + \nonumber \\ & \sum_{\substack{i\in S}}{\Delta^{sw}_i} \end{align}\]

Ref extensions

PowerModelsONM.ref_add_load_blocks! — Method

ref_add_load_blocks!(ref::Dict{Symbol,<:Any}, data::Dict{String,<:Any})

Ref extension to add load blocks to ref