Variables and Constraints

Variables

PowerModelsONM.variable_mc_block_indicator — Method

variable_mc_block_indicator(pm::AbstractUnbalancedPowerModel; nw::Int=nw_id_default, relax::Bool=false, report::Bool=true)

create variables for block status by load block

PowerModelsONM.variable_mc_switch_fixed — Method

variable_mc_switch_fixed(pm::AbstractUnbalancedPowerModel; nw::Int=nw_id_default, report::Bool=false)

Fixed switches set to constant values for multinetwork formulation (we need all switches)

PowerModelsONM.variable_mc_switch_state — Method

switch state (open/close) variables

Constraints

PowerModelsONM.constraint_block_isolation — Method

constraint_block_isolation(pm::AbstractUnbalancedPowerModel; nw::Int=nw_id_default, relax::Bool=false)

constraint to ensure that blocks are properly isolated by open switches

PowerModelsONM.constraint_mc_transformer_power_on_off — Method

constraint_mc_transformer_power(pm::AbstractUnbalancedPowerModel, i::Int; nw::Int=nw_id_default, fix_taps::Bool=true)::Nothing

Template function for Transformer constraints in Power-voltage space, considering winding type, conductor order, polarity and tap settings.

PowerModelsONM.constraint_storage_complementarity_mi_on_off — Method

constraint_storage_complementarity_mi_on_off(pm::AbstractUnbalancedPowerModel, i::Int; nw::Int=nw_id_default)

Template function for mixed-integer storage complementarity constraints

PowerModelsONM.constraint_switch_state_max_actions — Method

constraint_switch_state_max_actions(pm::AbstractUnbalancedPowerModel; nw::Int=nw_id_default)

max switching actions per timestep constraint

PowerModelsONM.constraint_block_isolation — Method

constraint_block_isolation(pm::PMD.LPUBFDiagModel, nw::Int; relax::Bool=false)

constraint to ensure that blocks get properly isolated by open switches by comparing the states of two neighboring blocks. If the neighboring block indicators are not either both 0 or both 1, the switch between them should be OPEN (0)

\[\begin{align} & (z^{bl}_{fr}_{i} - z^{bl}_{to}_{i}) \leq z^{sw}_{i}\ \forall i in S \\ & (z^{bl}_{fr}_{i} - z^{bl}_{fr}_{i}) \geq -z^{sw}_{i}\ \forall i in S \end{align} where $$z^{bl}_{fr}_{i}$$ and $$z^{bl}_{to}_{i}$$ are the indicator variables for the blocks on either side of switch $$i$$.\]

PowerModelsONM.constraint_switch_state_max_actions — Method

constraint_switch_state_max_actions(pm::AbstractUnbalancedPowerModel, nw_1::Int, nw_2::Int)

max actions per timestep switch constraint for multinetwork formulations

\[\sum_{\substack{i\in S}}{\Delta^{sw}_i} \leq z^{swu}\]

PowerModelsONM.constraint_switch_state_max_actions — Method

constraint_switch_state_max_actions(pm::AbstractUnbalancedPowerModel, nw::Int)

max actions per timestep switch constraint

\[\sum_{\substack{i\in S}}{\Delta^{sw}_i} \leq z^{swu}\]

PowerModelsONM.constraint_mc_storage_power_unbalance_mi — Method

constraint_mc_storage_power_unbalance_mi(pm::AbstractSwitchModels, nw::Int, i::Int, connections::Vector{Int}, balance_factor::Float64)

Enforces that storage inputs/outputs are (approximately) balanced across each phase, by some balance_factor

PowerModelsONM.constraint_storage_complementarity_mi_on_off — Method

constraint_storage_complementarity_mi_on_off(pm::LPUBFSwitchModel, n::Int, i::Int, charge_ub::Float64, discharge_ub::Float64)

scon + sdon == z_block

Objectives

PowerModelsONM.objective_mc_min_load_setpoint_delta_switch_global — Method

objective_mc_min_load_setpoint_delta_switch_global(pm::AbstractUnbalancedPowerModel)

minimum load delta objective without switch scores for global algorithm

\[\begin{align} \mbox{minimize: } & \nonumber \\ & \sum_{\substack{i\in N,c\in C}}{10 \left (1-z^v_i \right )} + \nonumber \\ & \sum_{\substack{i\in S}}{\Delta^{sw}_i} \end{align}\]

PowerModelsONM.objective_mc_min_load_setpoint_delta_switch_iterative — Method

objective_mc_min_load_setpoint_delta_switch_iterative(pm::AbstractUnbalancedPowerModel)

minimum load delta objective with switch scores for iterative algorithm

\[\begin{align} \mbox{minimize: } & \nonumber \\ & \sum_{\substack{i\in N,c\in C}}{10 \left (1-z^v_i \right )} + \nonumber \\ & \sum_{\substack{i\in L,c\in C}}{10 \omega_{i,c}\left |\Re{\left (S^d_i\right )}\right |\left ( 1-z^d_i \right ) } + \nonumber \\ & \sum_{\substack{i\in S}}{\Delta^{sw}_i} \end{align}\]

PowerModelsONM.objective_mc_min_storage_utilization — Method

objective_mc_min_storage_utilization(pm::AbstractUnbalancedPowerModel)

Minimizes the amount of storage that gets utilized in favor of using all available generation first

Ref extensions

PowerModelsONM.ref_add_load_blocks! — Method

ref_add_load_blocks!(ref::Dict{Symbol,<:Any}, data::Dict{String,<:Any})

Ref extension to add load blocks to ref

PowerModelsONM.ref_add_max_switch_actions! — Method

ref_add_max_switch_actions!(ref::Dict{Symbol,<:Any}, data::Dict{String,<:Any})

Ref extension to add maxswitchactions to ref, and set to Inf if option is missing